First commit
This is a strategy game with turn based tactical combat. It is inspired
by the xcom series of video games.

This first commit adds the basic components of the tactical gameplay:

A graph that controls where objects may move
Pathfinding
  In-editor visualization of graph
  In-editor tools to generate and modify the graph
  Snapping functionality to quickly add elements to the graph
  In-game selection tool to click on grid nodes.
Entities that can move around the graph, which have
  Traits which give "attribute" points,
  Triggers that traits can hook into, like the beginning of the turn,
    taking damage, etc.
  A streamlined "inventory" that connects types of items to an integer.
    This inventory is for both abstract and real items ("action points"
    and grenades, for example)
  Abilities, which consume action points and are activated in combat.
    For example, moving, shooting, etc.
The beginnings of a user interface for the tactical gameplay
  A menu showing all the selected unit's abilities as clickable squares
A simple turn system that schedules subclasses of type "Turn"
  "Squad" a Turn subclass that handles firing the BeginTurn and EndTurn
    events on the units that belong to it.
Animated characters that move smoothly along paths through the graph
Bezier curve visualization of paths.