Attack, Status, Body, Wound
This commit adds in the foundation for attacking. And because attacking causes injuries, it also introduces Body and Status, as well as a rework of Attributes, to enable interesting forms of damage.

Status is fairly simple, being just a plain Trait (beyond some syntactic sugar). Its function lies in how other Traits use it. Its basically a shared resource for all "temporary" changes to attributes. If you take health damage, a Trait can say Status.Lose(Attribute.Vitality, 3). Because Status is visible to all traits, another trait can heal that damage. Without such a "public" trait, damage would have to be "healed" by some form of tight coupling between the healing ability and the Traits that it specifically interacts with.

Attributes have been reworked to to better support how the current design of the game, in particular, the addition of "maladies". Maladies are negative traits. Positive values on maladies hurt the performance of the Entity that has them. Most maladies naturally heal (go down) over time. The maladies are "exhaustion" (wearing down of the body), "anguish" (wearing down of the mind), and "terror" and "arousal". The last three all relate to the mind, but whereas anguish generally affect all willpower rolls, terror and arousal are more narrow and their gameplay is based around spontaneous, random loss of control (e.g. soldier runs away in a panic). The other attributes are now "vitality" (Physical strength, replacing "health". May act as a "strength" stat in certain rolls, e.g. melee attacks), "willpower" (mental strength), "dexterity" (aim), "dodge" (lowers chance to be hit), "speed" (movement distance), "energy" (replacing ActionPoints), and "vigilance" (replacing ReactionPoints).

Body contains BodyParts, which for flavor (currently) may provide base stats and abilities (Chest provides vitality, legs give the Move ability, etc), but their function is be a home for Wounds and Disabilities. BodyParts have a score that represents how well they are working. Wounds may lower this score, and Disabilities respond to this score to lower the attributes of the Entity. For example, if and Entity's legs are reduce to 3/10 function, then the "injured" and "fractured" disabilities may both activate (the score being below both breakpoints), lowering the Entity's speed significantly, while also increasing anguish (which generally represents pain). Wounds, in addition to coming with attribute changes, like all traits, also have a "Bleed" property which allows them to cause attribute changes to Status at the end of every turn. Finally, BodyParts also have a similar feature to "Bleed" called "Recovery" which enables the same effect, but at the start of the turn. "Recovery" is intended to implement attribute healing, such as lowering exhaustion or terror. BodyParts that are lowered to 0 or less function are "crippled" and lose their recovery feature.

Spell is the base class for most interactive Abilities (Abilities that target Entities). It defines a "Caster" and "Target", as well as a "HitChance". When used, it rolls to see whether it was successful, and calls Succeed() or Fail() depending on the result.

Attack is a subclass of Spell, which adds in formulas to determine the HitChance, and the ability to apply Status damage and Wounds on Succeed() and Fail(). Attacks have two containers inside them, one of hit Wounds and one for miss Wounds (for the sake of generality, wounds and damage are allowed on misses). You can put as many Wounds as you want inside these containers, and Attack will select on of the them at random (you can control the odds by giving Wounds the "Weighted" script).

RangedAttack is a subclass of Attack that adds in to-hit bonuses based on the distance to the target (it does this through the MiscAttackBonus virtual property of Attack). It allows breaking up the range into 3 parts and applying bonuses depending on where the distance falls in that spectrum.

Energy (previously "ActionPoints") is no longer a purely "temporary" resource (meaning that you acquired it like currency), but instead you have intrinsic energy which you deplete (through Status) by performing actions. If you are missing energy, you may regain it through certain BodyParts that offer energy healing. Vigilance (previously "ReactionPoints") as well as all other item-like attributes are intended to work similarly.

Inventory has been dropped, instead being merged into Attributes. This allows all the powerful tools that work on Attributes to work on the elements that previously belonged to Inventory.

I made The MapCursor no longer a shared resource, instead expecting everything that uses a cursor to provide its own. There is, however, a DefaultCursor which turns on when no other cursor is active.

Added in Entity which is basically just a package that includes Attributes, Traits, and Triggers. Its convenient for things like targeting, where Critter is not general enough, and no other class is appropriate.